fix exit(1) at wrong location
Previously, I had persistent errors during the connectivity checks of crash report writer where the logs would reveal that a connectivity check has started but never finished. Now, I have discovered that there is this exit(1) that is located at the wrong place. It suggests that it was the cause of the failure of the connectivity check and perhaps Wails concealed the exit as it seems that the go code called from the frontend does not have the Wails go backend as parent and is not connected to the exit/panic logic of the Wails backend.